Move brew/cask stuff to Brewfile

+++ sublime/install.sh
@@ -1,15 +1,20 @@
-#!/bin/bash -ex
+#!/bin/bash -eu
+
+if [[ ! -d ~/Dropbox/Sublime/User ]] ; then
+  echo Failed to find ~/Dropbox/Sublime/User
+  exit 1
+fi
 
 SETTINGS_DIR="$HOME/Library/Application Support/Sublime Text 3"
 
-brew cask install sublime-text-dev
-brew install shellcheck
-
 mkdir -p "$SETTINGS_DIR/Packages/" "$SETTINGS_DIR/Installed Packages/"
 cd "$SETTINGS_DIR/Installed Packages/"
-wget https://sublime.wbond.net/Package%20Control.sublime-package
+
+[[ ! -f "Package Control.sublime-package" ]] && {
+  wget https://sublime.wbond.net/Package%20Control.sublime-package
+}
 
 cd "$SETTINGS_DIR/Packages/"
 [[ -d User ]] && rm -r User
 
-ln -fs "~/Dropbox (Personal)/Sublime/User"
+ln -fs ~/Dropbox/Sublime/User User
